server/requirements.txt
Flask==2.1.3
google-generativeai==0.1.0
